GatheringText.cpp -> LH3DText.cpp
