+++ config/BW1W120/splits.txt
@@ -1629,7 +1629,7 @@ Lionhead/LHLib/ver5.0/LHCollide.cpp:
 Lionhead/LH3DLib/development/LH3DRender.cpp:
 	.text       start:0x0082B220 end:0x00830050
 
-Black/GatheringText.cpp:
+Lionhead/LH3DLib/development/LH3DText.cpp:
 	.text       start:0x00830050 end:0x00833A00
 
 Lionhead/LH3DLib/development/LH3DWay.cpp:
