+++ configure.py
@@ -517,7 +517,6 @@ config.libs = [
             Object(NonMatching, "Black/GameThingWithPos.cpp"),
             Object(NonMatching, "Black/GatheringBox.cpp"),
             Object(NonMatching, "Black/GatheringInterface.cpp"),
-            Object(NonMatching, "Black/GatheringText.cpp"),
             Object(NonMatching, "Black/GenerateCreatureHTML.cpp"),
             Object(NonMatching, "Black/GestureSystem.cpp"),
             Object(NonMatching, "Black/GestureSystemData.cpp"),
@@ -1278,6 +1277,7 @@ config.libs = [
             Object(NonMatching, "Lionhead/LH3DLib/development/LH3DTech.cpp"),
             Object(NonMatching, "Lionhead/LHLib/ver5.0/LHCollide.cpp"),
             Object(NonMatching, "Lionhead/LH3DLib/development/LH3DRender.cpp"),
+            Object(NonMatching, "Lionhead/LH3DLib/development/LH3DText.cpp"),
             Object(NonMatching, "Lionhead/LH3DLib/development/LH3DWay.cpp"),
             Object(NonMatching, "Lionhead/LH3DLib/development/LH3DHair.cpp"),
         ],
